requirements.txt
flask
supabase
google-generativeai
python-dotenv
gunicorn
